refactor: move new ControlsBar into global components folder
